+++ packages/contracts/src/http/response.ts
@@ -1,5 +1,6 @@
 'use strict'
 
+import { ViewConfigBuilder } from '../view'
 import { CookieOptions } from './cookie-options'
 import { InteractsWithState } from './concerns/interacts-with-state'
 
@@ -175,10 +176,17 @@ export interface HttpResponse extends InteractsWithState {
    * @example
    * ```
    * response.view('welcome')
+   * response.view('welcome', view => {
+   *   view.layout('landing')
+   * })
    * response.view('user/dashboard', { user: { id: 1, name: 'Marcus' } })
+   * response.view('user/dashboard', { user: { id: 1, name: 'Marcus' } }, view => {
+   *   view.layout('profile')
+   * })
    * ```
    */
-  view (template: string, data?: any): Promise<this>
+  view (template: string, callback?: (viewBuilder: ViewConfigBuilder) => unknown): Promise<this>
+  view (template: string, data?: any, callback?: (viewBuilder: ViewConfigBuilder) => unknown): Promise<this>
 
   /**
    * Abort the request and throw an error with the given `status`. The status defaults