+++ packages/contracts/src/view/index.ts
@@ -1,3 +1,5 @@
 'use strict'
 
+export * from './config'
+export * from './config-builder'
 export * from './engine'